Vuejs

+++ API/src/dao/bookDAO.js
@@ -3,24 +3,77 @@ const prisma = new PrismaClient();
 const AppError = require('../model/AppError');
 const Book = require('../model/Book');
 
+function formatDate(date) {
+    return date ? new Date(date).toISOString().split('T')[0] : null;
+}
+
 class BookDAO {
 
     /**
-     * Récupère un livre par ID
-     * @returns {Promise<Object|null>} Liste de livre
-     * @throws {AppError} Si une erreur se produit lors de la récupération du livre
+     * Récupère tous les livres
+     * @returns {Promise<Array>} Liste de livres
+     * @throws {AppError} Si une erreur se produit lors de la récupération des livres
      */
     async getAll() {
         try {
-            return await prisma.book.findMany({
-                include: { // Inclure les catégories dans la réponse
+            const books = await prisma.book.findMany({
+                include: {
+                    author: {
+                        select: {
+                            id: true,
+                            name: true,
+                        },
+                    },
                     categories: true
                 }
             });
+            return books.map(book => ({
+                ...book,
+                publicationDate: formatDate(book.publicationDate)
+            }));
         } catch (error) {
-            throw new AppError('Erreur lors de la récupération du livre', 500);
+            this.handlePrismaError(error, 'Erreur lors de la récupération des livres');
         }
     }
+
+    /**
+     * Récupère un livre par ID
+     * @param {number} id
+     * @returns {Promise<Object|null>} Le livre trouvé ou null
+     * @throws {AppError} Si une erreur se produit lors de la récupération du livre
+     */
+    async getById(id) {
+        try {
+            id = parseInt(id);
+            if (isNaN(id)) {
+                throw new AppError('ID du livre invalide.', 400);
+            }
+
+            const book = await prisma.book.findUnique({
+                where: {
+                    id: id
+                },
+                include: {
+                    author: {
+                        select: {
+                            id: true,
+                            name: true,
+                        },
+                    },
+                    categories: true
+                }
+            });
+
+            if (book) {
+                book.publicationDate = formatDate(book.publicationDate);
+            }
+
+            return book;
+        } catch (error) {
+            this.handlePrismaError(error, 'Erreur lors de la récupération du livre');
+        }
+    }
+
     /**
      * Crée un livre dans la base de données avec des catégories
      * @param {Book} book
@@ -29,16 +82,13 @@ class BookDAO {
      */
     async create(book) {
         try {
-            // Assurez-vous que bookData est une instance de Book
-            if (!(book instanceof Book)) {
+            if (!(book instanceof Book) || book.validate(false).length > 0) {
                 throw new AppError("Données de livre invalides.", 400);
             }
-
-            const categoryIds = book.categories; // Utiliser le getter "categories"
+            const categoryIds = book.categories;
 
             const bookJson = book.toJson(false);
 
-            // Vérifie si categoryIds est défini et s'il contient des éléments
             const categoriesConnect = categoryIds && Array.isArray(categoryIds) && categoryIds.length > 0
                 ? {
                     categories: {
@@ -47,16 +97,14 @@ class BookDAO {
                 }
                 : {};
 
-            // Création du livre dans la base de données avec Prisma
             return await prisma.book.create({
                 data: {
-                    ...bookJson, // Les autres données du livre
-                    ...categoriesConnect // Ajouter les catégories connectées si elles existent
+                    ...bookJson,
+                    ...categoriesConnect
                 }
             });
-
         } catch (error) {
-            throw new AppError('Erreur lors de la création du livre', 400);
+            this.handlePrismaError(error, 'Erreur lors de la création du livre');
         }
     }
 
@@ -68,7 +116,10 @@ class BookDAO {
      */
     async update(book) {
         try {
-            // Vérifiez les catégories spécifiées
+            if (!(book instanceof Book) || book.validate(false).length > 0) {
+                throw new AppError("Données de livre invalides.", 400);
+            }
+
             const validCategories = await prisma.category.findMany({
                 where: { id: { in: book.categories } },
             });
@@ -80,8 +131,7 @@ class BookDAO {
                 );
             }
 
-            // Réattribuez les catégories et mettez à jour les autres champs
-            const updatedBook = await prisma.book.update({
+            return await prisma.book.update({
                 where: { id: book.id },
                 data: {
                     ...book.toJson(false),
@@ -91,12 +141,8 @@ class BookDAO {
                 },
             });
 
-            return updatedBook;
         } catch (error) {
-            throw new AppError(
-                "Erreur lors de la mise à jour du livre",
-                error.statusCode || 500
-            );
+            this.handlePrismaError(error, "Erreur lors de la mise à jour du livre");
         }
     }
 
@@ -110,58 +156,39 @@ class BookDAO {
         try {
             id = parseInt(id);
             if (isNaN(id)) {
-                throw new Error('ID du livre invalide');
+                throw new AppError('ID du livre invalide.', 400);
             }
             const book = await this.getById(id);
             if (!book) {
-                throw new AppError("Livre non trouvé", 400);
+                throw new AppError("Livre non trouvé", 404);
             }
-            return await prisma.book.deleteMany({
+
+            return await prisma.book.delete({
                 where: { id: id },
             });
 
         } catch (error) {
-            throw new Error('Erreur lors de la suppression du livre', error.message);
-        }
-    }
-
-    /**
-     * Récupère un livre par ID
-     * @param {number} id
-     * @returns {Promise<Object|null>} Le livre trouvé ou null
-     * @throws {AppError} Si une erreur se produit lors de la récupération du livre
-     */
-    async getById(id) {
-        try {
-            id = parseInt(id);
-            return await prisma.book.findUnique({
-                where: {
-                    id: id
-                },
-                include: { // Inclure les catégories dans la réponse
-                    categories: true
-                }
-            });
-        } catch (error) {
-            throw new AppError('Erreur lors de la récupération du livre', 500);
+            this.handlePrismaError(error, "Erreur lors de la suppression du livre");
         }
     }
 
     /**
      * Supprime des livres d'un auteur
      * @param {number} id
-     * @returns {Promise<Object>} Le livre supprimé
-     * @throws {AppError} Si une erreur se produit lors de la suppression du livre
+     * @returns {Promise<Object>} Le nombre de livres supprimés
+     * @throws {AppError} Si une erreur se produit lors de la suppression des livres
      */
     async deleteByAuthor(id) {
         try {
             id = parseInt(id);
-            const result = await prisma.book.deleteMany({
+            if (isNaN(id)) {
+                throw new AppError('ID de l\'auteur invalide.', 400);
+            }
+            return await prisma.book.deleteMany({
                 where: { authorId: id },
             });
-            return result.count;
         } catch (error) {
-            throw new AppError("Erreur lors de la suppression des livres de l'auteur", 500);
+            this.handlePrismaError(error, "Erreur lors de la suppression des livres de l'auteur");
         }
     }
 
@@ -169,20 +196,16 @@ class BookDAO {
      * Supprime tous les livres associés à une catégorie donnée.
      * @param {number} id - L'ID de la catégorie dont les livres doivent être supprimés.
      * @returns {Promise<void>}
-     * @throws {Error} Si une erreur se produit lors de la suppression des livres.
+     * @throws {AppError} Si une erreur se produit lors de la suppression des livres.
      */
     async deleteByCategory(id) {
         try {
             id = parseInt(id);
-            // Dissocier les livres de la catégorie
-            await prisma.bookCategory.deleteMany({
-                where: {
-                    categoryId: id,
-                },
-            });
+            if (isNaN(id)) {
+                throw new AppError('ID de la catégorie invalide.', 400);
+            }
 
-            // Supprimer les livres associés à la catégorie
-            await prisma.book.deleteMany({
+            return await prisma.book.deleteMany({
                 where: {
                     categories: {
                         some: {
@@ -192,20 +215,7 @@ class BookDAO {
                 },
             });
         } catch (error) {
-            throw new AppError("Erreur lors de la suppression des livres de la catégorie", 500);
-        }
-    }
-
-    async getBooksByAuthor(authorId) {
-        try {
-            authorId = parseInt(authorId);
-            return await prisma.book.findMany({
-                where: {
-                    authorId: authorId,
-                },
-            });
-        } catch (error) {
-            throw new AppError('Erreur lors de la récupération des livres de l\'auteur', 500);
+            this.handlePrismaError(error, "Erreur lors de la suppression des livres de la catégorie");
         }
     }
 
@@ -218,20 +228,68 @@ class BookDAO {
     async getBooksByCategory(categoryId) {
         try {
             categoryId = parseInt(categoryId);
+            if (isNaN(categoryId)) {
+                throw new AppError('ID de la catégorie invalide.', 400);
+            }
+
             return await prisma.book.findMany({
-                where: {
-                    categories: {
-                        some: {
-                            id: categoryId,
+                include: {
+                    author: {
+                        select: {
+                            id: true,
+                            name: true,
                         },
                     },
+                    categories: true
+                },
+                where: { categories: { some: { id: categoryId } } },
+            });
+        } catch (error) {
+            this.handlePrismaError(error, 'Erreur lors de la récupération des livres de la catégorie.');
+        }
+    }
+    async getBooksByAuthor(authorId) {
+        try {
+            authorId = parseInt(authorId);
+            return await prisma.book.findMany({
+                include: {
+                    author: {
+                        select: {
+                            id: true,
+                            name: true,
+                        },
+                    },
+                    categories: true
+                },
+                where: {
+                    authorId: authorId,
                 },
             });
         } catch (error) {
-            throw new AppError('Erreur lors de la récupération des livres de la catégorie', 500);
+            this.handlePrismaError(error, 'Erreur lors de la récupération des livres de l\'auteur.');
         }
     }
-
+    /**
+     * Gère les erreurs Prisma et les convertit en erreurs métiers.
+     * @param {Error} error - L'erreur levée par Prisma.
+     * @param {string} defaultMessage - Message par défaut pour les erreurs non spécifiées.
+     * @throws {AppError} - Une erreur personnalisée basée sur le contexte métier.
+     */
+    handlePrismaError(error, defaultMessage) {
+        if (error instanceof Prisma.PrismaClientKnownRequestError) {
+            switch (error.code) {
+                case 'P2002':
+                    throw new AppError("Une entrée avec ces données existe déjà.", 409);
+                case 'P2025':
+                    throw new AppError("Aucune correspondance trouvée pour cette requête.", 404);
+                case 'P2003':
+                    throw new AppError("Violation de contrainte de clé étrangère.", 400);
+                default:
+                    throw new AppError(defaultMessage, 500);
+            }
+        }
+        throw new AppError(defaultMessage, 400);
+    }
 }
 
 module.exports = new BookDAO();
